Extract `GradingControls` component from `GradingToolbar`
